refactor: app performance optimization
